Rename a few modules/functions to be more readable
